Update upscale_server.py
There was so much added for stability that this should be considered the first iteration of the working script for all practical purposes.  
Namely:
Automatically determined batch size and processing of that batch.
Automatically writes to a .json file the script can read from to resume processing on a video based upon the previously completed batch.
Fault tolerance if the GPU driver is unable to be detected through WSL2 because of Windows overhead holding onto resources.
Pillow based tiling. ESRGAN based tiling was considered, but cost too much overhead for too little benefit over the Pillow method.
Automated mounting of a ram disk for intermediate frames in each batch, which then reassembles those batches to a physical disk so they can be resumed.